Added BACKUP command
Support for hot-backing up a node.
You can retreive and restore a snapshot of the database to a file using
the BACKUP command
> BACKUP

Or using an HTTP connection like such:
$ curl localhost:7481/backup -o backup.db

The backup file format is a series of commands which are stored as
RESP Arrays.

The command:
SET mykey 123

Is stored on disk as:
"*3\r\n$3\r\nSET\r\n$5\r\nmykey\r\n$3\r\n123\r\n"

To restore a system from a backup, issue each command to the leader. For
example, using the `nc` command you could execute:
$ cat backup.db | nc localhost 7481